% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitnessEvaluation.R
\name{FitStats}
\alias{FitStats}
\title{Evaluate the fitness of different models}
\usage{
FitStats(y, p, r = NULL, fits)
}
\arguments{
\item{y}{A vector of observed values}

\item{p}{a matrix of predicted values}

\item{r}{a vector of naive values}

\item{fits}{a vector of fit statistics, including RMSE, MAD, RMSLE, MAPE, and MEAPE}
}
\value{
A matrix where each column corresponds with one of the 
fit statistics, and each row corresponds to a model.
}
\description{
Calculate the fit statistics of models, given the predicted values and observaed values.
}
\note{
Users can choose which statistics to calculate.
}
\examples{
x <- seq(1, 100, by=1)
y <- seq(1, 100, by=1) +  rnorm(100, 0, 1)
p <- matrix(predict(lm(y ~ x)), ncol=1)
r <- rep(median(y), 100)
FitStats(y=y, p=p, r=r, fits=c("rmse", "mad", "mrae"))
}
\author{
Luwei YING
}

